Moved SimpleAnimation to root of parts
